devops_agent/requirements.txt
google-adk>=1.0.0
kubernetes>=31.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
pytest>=8.0.0
